DK method: data type set to double
Single precision was working but sometimes created some artefacts

+++ src/mfx/dsp/va/dkm/Flt.h
@@ -37,7 +37,7 @@ namespace dkm
 
 
 // float looks robust enough
-typedef float Flt;
+typedef double Flt;
